resume-screening-app/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.9
werkzeug==2.3.7
google-generativeai==0.3.2
joblib==1.3.2
